i386: added tss
It was going to be needed anyway. Three stacks are needed
for our case: a ring 0 stack (used by `syscall` instruction);
a general interrupt stack (used by all interrupts except for
the timer); and a timer interrupt stack. A stack only for the
timer is needed because it may mess up the stack after an iret
on a nested interrupt.

i386: r8 register is now saved and restored

An old issue, which existed since interrupts were introduced
to the kernel. That also meant r8 was not preserved after
context switches (so tecnically, the kernel needed to be built
with `-ffixed-r8`). That was not the case.

scheduler: mutex::Unlock now  erases all other threads from the
internal list

Another old issue, dating back to mutex's introduction. I still
need to backport this to sched::semaphore. Might aswell remove
mutexes from the kernel entirely, and just use semaphores with
2 slots instead.